apps/oncolife/apps/doctor-platform/doctor-api/requirements.txt
# =============================================================================
# Doctor API - Python Dependencies
# =============================================================================
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.12.0

# Authentication (AWS Cognito)
boto3>=1.34.0
python-jose[cryptography]>=3.3.0
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
pytz>=2024.1
email-validator>=2.1.0

# Rate Limiting
slowapi>=0.1.9

# System Monitoring
psutil>=5.9.0

# Development
python-multipart>=0.0.6
httpx>=0.25.0

# Testing (optional - install with: pip install -r requirements.txt)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0



